Move capnp::text::Reader to a wrapper of data::Reader
This changes the capnp::text::Reader from a &str to a wrapper around
data::Reader, which only holds bytes.
This has the benefits of:
- Moving the conversion to &str and the associated utf8 check only when
  necessary, which can be a performance improvement in some cases.
- Allowing to support broken Text fields that may not hold valid UTF-8.

This is also however a breaking change as it forces to use `text::from_utf8`
to actually cast the bytes to &str instead of directly having &str.
Follows the same convention as https://doc.rust-lang.org/std/str/fn.from_utf8.html

This commit only holds the capnp runtime changes.