Add codegen to expose getter as bytes for text fields
Exposes as getter as bytes for text fields in the codegen. This is a
secondary getter that is for now only populating anything for text
fields for readers. I have duplicated the method and trimmed it only
for texts instead of complexifying the already existent getter_text
function for the sake of clarity.

This also regenerates the corresponding rpc_capnp and schema_capnp rust
files.